Port: Add Gouraud shading, fix GTE macros and 64-bit issues in rendering
- Add per-vertex color interpolation (Gouraud shading) to software rasterizer
  with PSX-accurate texture modulation (texel * vertex_color / 128)
- Read per-vertex colors from shade pipeline POLY_GT4 packs for level geometry
  and from obj->rgbs for preshaded characters (DG_FLAG_PAINT)
- Fix gte_stsxy3 macros: GT3/GT4 now write at correct offsets (8,20,32)
  instead of consecutive (8,10,12) which corrupted UV/color fields
- Fix DG_PVECTOR: long→int so struct is 8 bytes on 64-bit (matches SVECTOR)
- Fix DG_WriteObjVerticesIndirect: use SPAD struct access instead of raw
  offsets 0x3F8/0x3FC which read wrong bytes for 64-bit pointers
- Stub DG_TransChanl sets pack tags non-zero so shade processes all faces
- Skip shade pack colors for DG_MODEL_INDIRECT models: shade.c dereferences
  r0/g0/b0 as 32-bit pointer (broken on 64-bit), use neutral colors instead
- Preserve semi-transparency bit through texture color modulation
- Clamp interpolated colors to prevent negative values near triangle edges

+++ port/libdg/libdg.h
@@ -38,8 +38,8 @@ typedef struct DG_RVECTOR
 
 typedef struct DG_PVECTOR
 {
-    long vxy;
-    long vz;
+    int vxy;    /* Must be 4 bytes (not long, which is 8 on 64-bit) */
+    int vz;     /* Struct must be 8 bytes = sizeof(SVECTOR) */
 } DG_PVECTOR;
 
 /*---------------------------------------------------------------------------*/
